added analysis code to compute expectation values of other operators in the Mulliken population approximation
Also added output for 0th time step quantum variables.  (Previously there was a bug, and output began after the first step was taken.)

+++ examples/ethylene_sacasscf/start_c2h4.py
@@ -94,7 +94,7 @@ sim_params = {
     # final simulation time
     "max_quantum_time": traj_params["maxtime"],
     # initial qm amplitudes
-    "qm_amplitudes": np.ones(1),
+    "qm_amplitudes": np.ones(1,dtype=np.complex128),
     # energy shift used in quantum propagation
     "qm_energy_shift": 77.6,
 }
